+++ src/language/Keywords.kt
@@ -8,6 +8,8 @@ val KEYWORDS = arrayOf(
         "or",
         "xor",
         "not",
+        "equal",
+        "equals",
         "mod"
         //TODO: Add Java keywords, or change any variable name that is a Java keyword
 )