+++ Changelog.md
@@ -4,6 +4,7 @@ In development
 * Added the ability to declare and assign variables
 * Declaring variables has the syntax `new [type] variable [name] = [value]` (constant can be written instead of variable to make it immutable)
 * Assigning variables has the same syntax as setting attributes: `set [name] to [value]`
+* Objects now contain fields which can be referenced and used (but not reassigned) within your code using the syntax `object's field`, for example: `player's position` is a vector equal to the player's current position in the world
 * Added several new data types: boolean, integer, decimal, string, vector and world (note that most of these do not yet have all of the desired features, such as an easy way to perform arithmetic)
 * All block behaviours now have access to the world
 * Added an attribute class to store the value and data type of the different attributes for items, blocks, etc